% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clock_data.R
\name{concatDesignRuns}
\alias{concatDesignRuns}
\title{Concatenate design matrices for each run to form a single design with unique baselines per run (ala AFNI)}
\usage{
concatDesignRuns(d)
}
\description{
Concatenate design matrices for each run to form a single design with unique baselines per run (ala AFNI)
}

