% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clock_model.R
\docType{class}
\name{deltavalue_model}
\alias{deltavalue_model}
\alias{deltavalue_model-class}
\title{Simple delta rule model for learning expected value across trials.}
\description{
Just uses the observed reward history to fit a single learning rate
representing the extent to which the agent discounts the information
from the reinforcement history.
}
\details{
The goal of this model is to provide a benchmark
}

