% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{detrendts}
\alias{detrendts}
\title{detrend time series up to quadratic trend. Used by fir1Bandpass prior to filtering}
\usage{
detrendts(x, order = 0)
}
\description{
detrend time series up to quadratic trend. Used by fir1Bandpass prior to filtering
}
\keyword{internal}

