% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clock_params.R
\name{getRTUpdate.default}
\alias{getRTUpdate.default}
\title{default RT prediction method for a parameter (not used)}
\usage{
\method{getRTUpdate}{default}(obj, theta)
}
\description{
default RT prediction method for a parameter (not used)
}
\keyword{internal}

