\name{getRTUpdate.p_go}
\alias{getRTUpdate.p_go}
\title{Compute reaction time contribution for go (speedup for PPE)}
\usage{
\method{getRTUpdate}{p_go}(obj, theta)
}
\arguments{
  \item{obj}{the parameter object to be used for
  prediction.}

  \item{theta}{named vector of current values for
  parameters in model.}
}
\description{
Compute reaction time contribution for go (speedup for PPE)
}
\keyword{internal}

