\name{getRTUpdate.p_nogo}
\alias{getRTUpdate.p_nogo}
\title{Compute reaction time contribution for noGo (slowdown for NPE)}
\usage{
\method{getRTUpdate}{p_nogo}(obj, theta)
}
\arguments{
  \item{obj}{the parameter object to be used for
  prediction.}

  \item{theta}{named vector of current values for
  parameters in model.}
}
\description{
Compute reaction time contribution for noGo (slowdown for
NPE)
}
\keyword{internal}

