% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clock_params.R
\name{getRTUpdate}
\alias{getRTUpdate}
\title{general dispatch of RT prediction for a parameter}
\usage{
getRTUpdate(obj, theta)
}
\description{
general dispatch of RT prediction for a parameter
}
\keyword{internal}

