% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTheta}
\alias{getTheta}
\title{internal function to lookup the correct value of theta when a parameter varies by condition}
\usage{
getTheta(obj, condition)
}
\description{
internal function to lookup the correct value of theta when a parameter varies by condition
}
\keyword{internal}

