\name{reset_workspace.p_epsilonBeta}
\alias{reset_workspace.p_epsilonBeta}
\title{reset workspace for epsilonBeta parameter}
\usage{
\method{reset_workspace}{p_epsilonBeta}(obj)
}
\arguments{
  \item{obj}{the parameter object whose workspace will be
  reset.}
}
\description{
Allocates the beta distribution tracker if not yet
instantiated.
}

