% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clock_params.R
\name{reset_workspace.p_gold}
\alias{reset_workspace.p_gold}
\title{reset workspace for goForGold parameter}
\usage{
\method{reset_workspace}{p_gold}(obj)
}
\arguments{
\item{obj}{the parameter object whose workspace will be reset.}
}
\description{
Resets the bestRT for scale parameter
}

