% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clock_params.R
\name{reset_workspace.p_nogo}
\alias{reset_workspace.p_nogo}
\title{reset workspace for noGo parameter}
\usage{
\method{reset_workspace}{p_nogo}(obj)
}
\arguments{
\item{obj}{the parameter object whose workspace will be reset.}
}
\description{
Resets the NoGo term for scaling NPE slowdown
}

