\name{reset_workspace.p_stickyChoice}
\alias{reset_workspace.p_stickyChoice}
\title{reset workspace for stickyChoice parameter}
\usage{
\method{reset_workspace}{p_stickyChoice}(obj)
}
\arguments{
  \item{obj}{the parameter object whose workspace will be
  reset.}
}
\description{
Reset the sticky choice function value
}

