% Generated by roxygen2 (4.0.2): do not edit by hand
\name{detrendts}
\alias{detrendts}
\title{detrend time series up to quadratic trend. Used by fir1Bandpass prior to filtering}
\usage{
detrendts(x, order = 0)
}
\description{
detrend time series up to quadratic trend. Used by fir1Bandpass prior to filtering
}
\keyword{internal}

