% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getRTUpdate.default}
\alias{getRTUpdate.default}
\title{default RT prediction method for a parameter (not used)}
\usage{
\method{getRTUpdate}{default}(obj, theta)
}
\description{
default RT prediction method for a parameter (not used)
}
\keyword{internal}

