\name{getRTUpdate.p_epsilonBeta}
\alias{getRTUpdate.p_epsilonBeta}
\title{Compute reaction time contribution for epsilon (explore)}
\usage{
\method{getRTUpdate}{p_epsilonBeta}(obj, theta)
}
\arguments{
  \item{obj}{the parameter object to be used for
  prediction.}

  \item{theta}{named vector of current values for
  parameters in model.}
}
\description{
Compute reaction time contribution for epsilon (explore)
}
\keyword{internal}

