\name{getRTUpdate.p_meanRT}
\alias{getRTUpdate.p_meanRT}
\title{Compute reaction time contribution for K (baseline response speed)}
\usage{
\method{getRTUpdate}{p_meanRT}(obj, theta)
}
\arguments{
  \item{obj}{the parameter object to be used for
  prediction.}

  \item{theta}{named vector of current values for
  parameters in model.}
}
\description{
Compute reaction time contribution for K (baseline response
speed)
}
\keyword{internal}

