% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clock_params.R
\name{getTheta}
\alias{getTheta}
\title{internal function to lookup the correct value of theta when a parameter varies by condition}
\usage{
getTheta(obj, condition)
}
\description{
internal function to lookup the correct value of theta when a parameter varies by condition
}
\keyword{internal}

