\name{reset_workspace.p_meanSlowFast}
\alias{reset_workspace.p_meanSlowFast}
\title{reset workspace for rho parameter}
\usage{
\method{reset_workspace}{p_meanSlowFast}(obj)
}
\arguments{
  \item{obj}{the parameter object whose workspace will be
  reset.}
}
\description{
Allocates the beta distribution tracker if not yet
instantiated.
}

