\name{reset_workspace}
\alias{reset_workspace}
\title{Generic dispatch to reset relevant variables in the workspace for a parameter.}
\usage{
reset_workspace(obj)
}
\arguments{
  \item{obj}{the parameter object whose workspace will be
  reset.}
}
\description{
Generic dispatch to reset relevant variables in the
workspace for a parameter.
}

