% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runmean}
\alias{runmean}
\title{compute a moving average smooth over a time series (here, a vector of RTs)}
\usage{
runmean(x, k = 5)
}
\description{
used to fit smoothed RTs (\code{clock_model} object).
Function is an adapted version of \code{runmean} from the \code{caTools} package.
}
\keyword{internal}

