\name{updateBetaDists}
\alias{updateBetaDists}
\title{updates counts of beta distributions for fast and slow responses
when there is a PPE in the fast versus slow direction.}
\usage{
updateBetaDists(bfs)
}
\arguments{
  \item{bfs}{betaFastSlow object containing fast and slow
  beta dist parameters}
}
\description{
updates counts of beta distributions for fast and slow
responses when there is a PPE in the fast versus slow
direction.
}
\keyword{internal}

